Create 'FeedStoreSpecs' that must be implemented by any 'FeedStore' test implementation to guarantee it meets specs
